Modifier Type Repository

+++ SOSAuto_BACKEND/src/main/java/com/sosauto_backend/respository/AdministrateurRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdministrateurRepository extends JpaRepository<Administrateur, Integer> {
+public interface AdministrateurRepository extends JpaRepository<Administrateur, Long> {
 
 }
